requirements.txt
absl-py==0.8.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
flask-heroku==0.1.9
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jws==0.1.3
numpy==1.17.2
numpydoc==0.9.1
Pyrebase4==4.3.0
scikit-learn==0.21.3
SQLAlchemy==1.3.9